refactor: introduce hyperswitch_interface crate
