--- crates/encryption_interface/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "encryption_interface"
-version = "0.1.0"
-edition.workspace = true
-rust-version.workspace = true
-readme = "README.md"
-license.workspace = true
-
-[dependencies]
-async-trait = "0.1.68"
-dyn-clone = "1.0.11"
-thiserror = "1.0.40"
-
-# First party crates
-common_utils = { version = "0.1.0", path = "../common_utils" }
