--- crates/encryption_interface/README.md
@@ -1,4 +1,0 @@
-# Encryption Interface
-
-This crate includes encryption interface and its error types which are implemented
-for the supporting kms clients
