+++ crates/external_services/Cargo.toml
@@ -35,7 +35,6 @@ hex = "0.4.3"
 
 # First party crates
 common_utils = { version = "0.1.0", path = "../common_utils" }
-encryption_interface = { version = "0.1.0", path = "../encryption_interface" }
+hyperswitch_interface = { version = "0.1.0", path = "../hyperswitch_interface" }
 masking = { version = "0.1.0", path = "../masking" }
 router_env = { version = "0.1.0", path = "../router_env", features = ["log_extra_implicit_fields", "log_custom_entries_to_extra"] }
-secrets_interface = { version = "0.1.0", path = "../secrets_interface" }
