+++ crates/external_services/src/managers.rs
@@ -1,0 +1,5 @@
+//! Config and client managers
+
+pub mod encryption_management;
+
+pub mod secrets_management;
