+++ crates/hyperswitch_interface/README.md
@@ -1,0 +1,3 @@
+# Hyperswitch Interface
+
+This crate includes interfaces and its error types
