+++ crates/hyperswitch_interface/src/lib.rs
@@ -1,0 +1,7 @@
+//! Hyperswitch interface
+
+#![warn(missing_docs, missing_debug_implementations)]
+
+pub mod secrets_interface;
+
+pub mod encryption_interface;
