--- crates/secrets_interface/README.md
@@ -1,4 +1,0 @@
-# Secrets Interface
-
-This crate includes secrets management interface and its error types which are implemented
-for the supporting kms clients
